+++ src/Stride.Shaders.Compilers/Direct3D/FXC.cs
@@ -2,7 +2,7 @@ using System.Text;
 using Silk.NET.Core.Native;
 using Silk.NET.Direct3D.Compilers;
 
-namespace Stride.Shaders.Compilers;
+namespace Stride.Shaders.Compilers.Direct3D;
